make evaluator generic
